volume persistant

+++ docker/php/entrypoint.sh
@@ -7,7 +7,11 @@ if [ "$#" -gt 0 ]; then
 fi
 
 if [ ! -f /var/www/html/public/uploads/users/default-profile.png ]; then
-    cp /var/www/html/public/uploads/users/default-profile.png.bak /var/www/html/public/uploads/users/default-profile.png
+    if [ -f /var/www/html/public/uploads/users/default-profile.png.bak ]; then
+        cp /var/www/html/public/uploads/users/default-profile.png.bak /var/www/html/public/uploads/users/default-profile.png
+    else
+        echo "INFO: default-profile.png.bak not found, skipping default photo copy"
+    fi
 fi
 
 php-fpm &
